+++ app/containers/AgentDesktop/constants.js
@@ -104,8 +104,8 @@ export const SET_DISPOSITION_DETAILS =
   'app/AgentDesktop/SET_DISPOSITION_DETAILS';
 export const SELECT_DISPOSITION = 'app/AgentDesktop/SELECT_DISPOSITION';
 export const SHOW_REFRESH_NOTIF = 'app/AgentDesktop/SHOW_REFRESH_NOTIF';
-export const SHOW_CONTACTS_PANEL = 'app/AgentDesktop/SHOW_CONTACTS_PANEL';
-export const HIDE_CONTACTS_PANEL = 'app/AgentDesktop/HIDE_CONTACTS_PANEL';
+export const SHOW_SIDE_PANEL = 'app/AgentDesktop/SHOW_SIDE_PANEL';
+export const HIDE_SIDE_PANEL = 'app/AgentDesktop/HIDE_SIDE_PANEL';
 export const SHOW_INTERACTIONS_BAR = 'app/AgentDesktop/SHOW_INTERACTIONS_BAR';
 export const HIDE_INTERACTIONS_BAR = 'app/AgentDesktop/HIDE_INTERACTIONS_BAR';
 export const SET_FORM_IS_DIRTY = 'app/ContactsControl/SET_FORM_IS_DIRTY';