+++ app/containers/AgentDesktop/tests/selectors.test.js
@@ -74,6 +74,7 @@ describe('selectAwaitingDisposition', () => {
             },
           },
         ],
+        noInteractionContactPanel: {},
       },
     });
     expect(selectAwaitingDisposition(mockedState)).toEqual(true);